+++ mask_safe.go
@@ -7,7 +7,7 @@
 
 package websocket
 
-func maskBytes(key [4]byte, pos int, b []byte) int {
+func maskBytes(key [4]byte, pos uint8, b []byte) uint8 {
 	for i := range b {
 		b[i] ^= key[pos&3]
 		pos++
